requirements.txt
streamlit==1.32.0

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38

langchain-google-genai==1.0.10
google-generativeai==0.7.2

langchain-huggingface
sentence-transformers

faiss-cpu==1.8.0

PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.1
openpyxl==3.1.2

python-dotenv==1.0.1